Add support for Open Protocol in Device Connection
This commit includes the addition of "Open Protocol" in the communication protocol options, and defines connection details associated with it, which includes required host and port of the device under the new section "OpenProtocolConnDetails". This change was needed to allow devices that use "Open Protocol" to be seamlessly integrated into our system.